changes to ppl world etc
